take care of non pie binary or pie binary
