+++ libafl_qemu/src/modules/utils/mod.rs
@@ -1,1 +1,3 @@
+pub mod addr2line;
 pub mod filters;
+pub use addr2line::*;
